Import bug fixes

+++ Core/Domain/Entities/Fertilization.cs
@@ -21,9 +21,9 @@ namespace Rems.Domain.Entities
 
         public int? UnitId { get; set; }
 
-        public DateTime Date { get; set; }        
+        public DateTime Date { get; set; }
 
-        public int? Amount { get; set; }
+        public double? Amount { get; set; }
 
         public int? Depth { get; set; }
